Move key-generation service ownership to KM
