+++ libs/common/src/auth/services/token.service.spec.ts
@@ -9,12 +9,12 @@ import { LogoutReason } from "@bitwarden/auth/common";
 
 import { FakeSingleUserStateProvider, FakeGlobalStateProvider } from "../../../spec";
 import { EncryptService } from "../../key-management/crypto/abstractions/encrypt.service";
+import { KeyGenerationService } from "../../key-management/crypto/abstractions/key-generation.service";
 import {
   VaultTimeout,
   VaultTimeoutAction,
   VaultTimeoutStringType,
 } from "../../key-management/vault-timeout";
-import { KeyGenerationService } from "../../platform/abstractions/key-generation.service";
 import { LogService } from "../../platform/abstractions/log.service";
 import { AbstractStorageService } from "../../platform/abstractions/storage.service";
 import { StorageLocation } from "../../platform/enums";
