+++ libs/common/src/key-management/master-password/services/master-password.service.spec.ts
@@ -4,7 +4,6 @@ import * as rxjs from "rxjs";
 
 import { makeSymmetricCryptoKey } from "../../../../spec";
 import { ForceSetPasswordReason } from "../../../auth/models/domain/force-set-password-reason";
-import { KeyGenerationService } from "../../../platform/abstractions/key-generation.service";
 import { LogService } from "../../../platform/abstractions/log.service";
 import { StateService } from "../../../platform/abstractions/state.service";
 import { EncString } from "../../../platform/models/domain/enc-string";
@@ -13,6 +12,7 @@ import { StateProvider } from "../../../platform/state";
 import { UserId } from "../../../types/guid";
 import { MasterKey } from "../../../types/key";
 import { EncryptService } from "../../crypto/abstractions/encrypt.service";
+import { KeyGenerationService } from "../../crypto/abstractions/key-generation.service";
 
 import { MasterPasswordService } from "./master-password.service";
