+++ libs/common/src/tools/send/services/send.service.ts
@@ -7,8 +7,8 @@ import { Observable, concatMap, distinctUntilChanged, firstValueFrom, map } from
 import { PBKDF2KdfConfig, KeyService } from "@bitwarden/key-management";
 
 import { EncryptService } from "../../../key-management/crypto/abstractions/encrypt.service";
+import { KeyGenerationService } from "../../../key-management/crypto/abstractions/key-generation.service";
 import { I18nService } from "../../../platform/abstractions/i18n.service";
-import { KeyGenerationService } from "../../../platform/abstractions/key-generation.service";
 import { Utils } from "../../../platform/misc/utils";
 import { EncArrayBuffer } from "../../../platform/models/domain/enc-array-buffer";
 import { EncString } from "../../../platform/models/domain/enc-string";
